Build MacOS binaries on MacOS hosts
In order for DNS resolution to work on MacOS in some environments, go
projects must be built with on a MacOS host with `-tags=netcgo` and
CGO_ENABLED not explicity set to either 0 or 1.   This is the default
behavor when running `go build` on MacOS.

Explicitly enabling CGO with `CGO_ENABLED=1` causes issues when
cross-compiling between GOARCHs on MacOS, related to the version of the
MacOS SDK that is selected. This may not be an issue in future MacOS
releases. Disabling CGO with `CGO_ENABLED=0` causes `-tags=netcgo` to be
ignored, preventing DNS from functioning.

This change seperates the `GOOS=darwin` from the remainder of the
supported `GOOS` targets behind different goreleaser build ids. This
allows us to build the MacOS releases on MacOS hosts in CI, and the
remainder can be cross-compiled from a Linux host.

Running goreleaser outside of CI will remain unaffected by this change
with the exception of building for MacOS from non-MacOS hosts, in which
case the resulting binaries may exhibit DNS issues but should otherwise
function. Users can skip building for MacOS in this case by running
`goreleaser build --id=default`.

For reference, the upstream issue with more details about DNS resolution
on MacOS can be found in this GitHub issue: https://github.com/golang/go/issues/12524

Signed-off-by: Scott Macfarlane <smacfarlane@hashicorp.com>